fix auth tests

+++ src/comments/comments.service.ts
@@ -1,33 +1,71 @@
-import { Injectable } from '@nestjs/common';
+import {
+  BadRequestException,
+  Injectable,
+  NotFoundException,
+  UnprocessableEntityException,
+} from '@nestjs/common';
 import { CreateCommentDto } from './dto/create-comment.dto';
 import { PrismaService } from 'src/prisma/prisma.service';
-import { Prisma } from 'generated/prisma/client';
+import { Comment, Prisma } from 'generated/prisma/client';
+import { validate } from 'uuid';
 
 @Injectable()
 export class CommentsService {
+  transformComment(comment: Comment) {
+    return {
+      ...comment,
+      createdAt: comment.createdAt.getTime(),
+      updatedAt: comment.updatedAt.getTime(),
+    };
+  }
   constructor(private prisma: PrismaService) {}
 
-  create(data: CreateCommentDto) {
-    return this.prisma.comment.create({
+  async create(data: CreateCommentDto) {
+    if (!validate(data.articleId)) {
+      throw new BadRequestException('Invalid article ID');
+    }
+
+    const articleExists = await this.prisma.article.findUnique({
+      where: { id: data.articleId },
+    });
+    if (!articleExists) {
+      throw new UnprocessableEntityException('Article not found');
+    }
+
+    const comment = await this.prisma.comment.create({
       data: {
-        ...data,
-        authorId: undefined,
-        author: { connect: { id: data.authorId } },
-        articleId: undefined,
-        article: data.articleId ? { connect: { id: data.articleId } } : null,
+        content: data.content,
+        article: { connect: { id: data.articleId } },
+        author: data.authorId ? { connect: { id: data.authorId } } : undefined,
       } as Prisma.CommentCreateInput,
     });
+    return this.transformComment(comment);
   }
 
-  findAll(articleId?: string) {
-    return this.prisma.comment.findMany({
+  async findAll(articleId?: string) {
+    const comments = await this.prisma.comment.findMany({
       where: {
         articleId,
       },
     });
+    return comments.map((comment) => this.transformComment(comment));
   }
 
-  delete(where: Prisma.CommentWhereUniqueInput) {
+  async findOne(where: Prisma.CommentWhereUniqueInput) {
+    const comment = await this.prisma.comment.findFirst({
+      where,
+    });
+    if (!comment) {
+      throw new NotFoundException('Comment not found');
+    }
+    return this.transformComment(comment);
+  }
+
+  async delete(where: Prisma.CommentWhereUniqueInput) {
+    const comment = await this.prisma.comment.findUnique({
+      where,
+    });
+    if (!comment) throw new NotFoundException('Comment not found');
     return this.prisma.comment.delete({
       where,
     });
